+++ test/compojure/api/middleware_test.clj
@@ -3,7 +3,8 @@
             [midje.sweet :refer :all]
             [ring.util.http-response :refer [ok]]
             [ring.util.http-status :as status]
-            ring.util.test)
+            ring.util.test
+            [slingshot.slingshot :refer [throw+]])
   (:import [java.io PrintStream ByteArrayOutputStream]))
 
 (defmacro without-err
@@ -40,13 +41,34 @@
 
 (facts "wrap-exceptions"
   (with-out-str
-    (without-err
-      (let [exception (RuntimeException. "kosh")
-            exception-class (.getName (.getClass exception))
-            failure (fn [_] (throw exception))]
+   (without-err
+     (let [exception (RuntimeException. "kosh")
+           exception-class (.getName (.getClass exception))
+           handler (-> (fn [_] (throw exception))
+                       (wrap-exceptions {}))]
 
-        (fact "converts exceptions into safe internal server errors"
-          ((wrap-exceptions failure (:handlers (:exceptions api-middleware-defaults))) ..request..)
-          => (contains {:status status/internal-server-error
-                        :body (contains {:class exception-class
-                                         :type "unknown-exception"})}))))))
+       (fact "converts exceptions into safe internal server errors"
+         (handler {}) => (contains {:status status/internal-server-error
+                                    :body (contains {:class exception-class
+                                                     :type "unknown-exception"})})))))
+
+  (with-out-str
+   (without-err
+     (fact "Slingshot exception map type can be matched"
+       (let [handler (-> (fn [_] (throw+ {:type ::test} (RuntimeException. "kosh")))
+                         (wrap-exceptions {:handlers {::test (fn [ex _ _] {:status 500 :body "hello"})}}))]
+         (handler {}) => (contains {:status status/internal-server-error
+                                    :body "hello"})))))
+
+  (without-err
+    (fact "Default log-fn logs exceptions to console"
+      (let [handler (-> (fn [_] (throw (RuntimeException. "kosh")))
+                        (wrap-exceptions {}))]
+        (with-out-str (handler {})) => "ERROR kosh\n")))
+
+  (fact "Log-fn can be overriden"
+    (let [t (atom false)
+          handler (-> (fn [_] (throw (RuntimeException. "kosh")))
+                      (wrap-exceptions {:log-fn (fn [_] (reset! t true) nil)}))]
+      (with-out-str (handler {})) => ""
+      @t => true)))